Add logic to collect observable in DooD setup

+++ scripts/unwrap.sh
@@ -55,13 +55,22 @@ flock "$REFCOUNT_FILE" sh -c '
         echo "$SUDO"
         ls -hlia /usr/bin/runc
         ls -hlia /usr/bin/runc.bkp
-        $SUDO mv /usr/bin/runc.bkp /usr/bin/runc
+        if [ -f /usr/bin/runc.bkp ]; then
+            $SUDO mv /usr/bin/runc.bkp /usr/bin/runc
+        fi
         ls -hlia /usr/bin/runc
 
         ls -hlia  /mnt/curiosity/docker-buildx.bkp
         ls -hlia  /usr/libexec/docker/cli-plugins/docker-buildx
-        $SUDO mv /mnt/curiosity/docker-buildx.bkp /usr/libexec/docker/cli-plugins/docker-buildx
+        if [ -f /mnt/curiosity/docker-buildx.bkp ]; then
+            $SUDO mv /mnt/curiosity/docker-buildx.bkp /usr/libexec/docker/cli-plugins/docker-buildx
+        fi
         ls -hlia  /usr/libexec/docker/cli-plugins/docker-buildx
         echo "OK: Done unwrapping"
+
+        docker ps
+        docker stop co-docker > /dev/null 2>&1 && docker rm co-docker > /dev/null 2>&1
+        docker ps
+        echo "OK: Done stopping and removing co-docker"
     fi
 ' _ "$REFCOUNT_FILE"
